+++ CHANGES.rst
@@ -1,2 +1,13 @@
-2.0.0 (2016-07-28)
+1.1.0
+----------
+
+- Allow redirect to for all external links on any page.
+- Add RegEx url match for links on any site to be redirected.
+
+
+1.0.0 
 ------------------
+
+- Create a custom redirect/confirmation page.
+- Allow redirection for external links on any page.
+- Allow redirection for internal links on any page.